research/requirements.txt
pandas>=1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
openpyxl>=3.0.0
